chore: added mocks folder to separate mock components
